Refactor : 폴더 및 불필요 파일 수정
