+++ src/Router.jsx
@@ -24,7 +24,7 @@ const Search = retryLazy(() => import("./pages/Search"));
 const NotFoundScene = retryLazy(() => import("./pages/NotFoundScene"));
 const BoardEditPost = retryLazy(() => import("./pages/BoardEditPost"));
 const ThreeBoardEditPost = retryLazy(() =>
-  import("./pages/ThreeBoardEditPost")
+  import("./components/Slider/ThreeBoardEditPost")
 );
 const Board = retryLazy(() => import("./pages/Board"));
 const BoardPost = retryLazy(() => import("./pages/BoardPost"));
