+++ src/components/Slider/CategoryCommercial.jsx
@@ -1,5 +1,5 @@
 import styled, { css } from "styled-components";
-import SliderSingleMain from "./Slider/SliderSingleMain";
+import SliderSingleMain from "./SliderCore/SliderSingleMain";
 import youtubeData from "./data/youtubeData";
 import healthData from "./data/healthData";
 import commercialData from "./data/commercialData";
