+++ src/components/Slider/SliderCore/SliderMultiItem.jsx
