+++ src/components/Slider/SliderCore/SliderMultiMain.jsx
