+++ src/components/Slider/SliderCore/SliderSingleMain.jsx
