+++ src/components/User/UserLogin.jsx
@@ -9,7 +9,7 @@ import {
   LinkGroup,
 } from "./LoginLayout";
 import axios from "axios";
-import Button from "../common/Button";
+import Button from "../../share/Button";
 import useUserData from "../../api/useUserData";
 
 export default function UserInputForm() {
