--- src/utils/Editor/Details.jsx
@@ -1,11 +1,0 @@
-import parse from "html-react-parser";
-
-const Details = ({ description }) => {
-    return (
-        <>
-            <div className="ProseMirror">{parse(description)}</div>
-        </>
-    );
-};
-
-export default Details;
